+++ rails/db/migrate/20240628174726_create_concepts.rb
@@ -3,10 +3,10 @@ class CreateConcepts < ActiveRecord::Migration[7.1]
     create_table :concepts do |t|
       t.string :wikidata
       t.string :display_name
-      t.integer :level
+      t.bigint :level
       t.string :description
-      t.integer :works_count
-      t.integer :cited_by_count
+      t.bigint :works_count
+      t.bigint :cited_by_count
       t.string :image_url
       t.string :image_thumbnail_url
       t.string :works_api_url
