+++ rails/db/migrate/20240628174739_create_sources.rb
@@ -5,8 +5,8 @@ class CreateSources < ActiveRecord::Migration[7.1]
       t.jsonb :issn
       t.string :display_name
       t.string :publisher
-      t.integer :works_count
-      t.integer :cited_by_count
+      t.bigint :works_count
+      t.bigint :cited_by_count
       t.boolean :is_oa
       t.boolean :is_in_doaj
       t.string :homepage_url
